♻️ refactor : 폴더 구조 개선과 컴포넌트 추가 #73
